Split CommentItem into CommentItem and CommentsItemMeta
