+++ packages/lesswrong/components/comments/SingleLineComment.tsx
@@ -5,9 +5,9 @@ import classNames from 'classnames';
 import withErrorBoundary from '../common/withErrorBoundary';
 import { commentGetKarma } from '../../lib/collections/comments/helpers'
 import { isMobile } from '../../lib/utils/isMobile'
-import { styles as commentsItemStyles } from './CommentsItem/CommentsItem';
 import { CommentTreeOptions } from './commentTree';
 import { coreTagIconMap } from '../tagging/CoreTagIcon';
+import { metaNoticeStyles } from './CommentsItem/CommentsItemMeta';
 
 export const SINGLE_LINE_PADDING_TOP = 5
 
@@ -132,11 +132,11 @@ const styles = (theme: ThemeType): JssStyles => ({
     }
   },
   metaNotice: {
-    ...commentsItemStyles(theme).metaNotice,
+    ...metaNoticeStyles(theme),
     marginRight: theme.spacing.unit
   },
   postTitle: {
-    ...commentsItemStyles(theme).metaNotice,
+    ...metaNoticeStyles(theme),
     marginRight: 20
   },
   preview: {
